requirements.txt
Flask>=3.1.1
youtube-transcript-api>=1.0.3
python-dotenv>=1.1.0
google-genai>=1.17.0